Add docker network alias to containers
By having the alias of the service to be the same as its FQDN, other containers
on the same docker network can access this service without an external DNS
lookup. This is most helpful during staging, where we don't want services
communicating with the live servers but instead with other staging services.

This was happening with some services reaching out to the production Gitlab
server when they needed to reach the staging version.

Mantis: 5101